Add digest for Strategy team to track progress of chartering issues (#294)
Goal is to send an email to the Strategy team on Monday when there have been
updates in charter issues in the Strategy pipeline to propose the addition of
an agenda item to review the charters.

A specific digest template gets used to create a more meaningful email subject
as well as to link to the charter in progress dashboard.

It may turn out that it would be better to just report the creation of an
issue with a "charter" label whenever such an issue gets created. To be
assessed after a few weeks...